@@ -1,22 +1,16 @@
 import 'reflect-metadata';
-import 'rxjs/Rx';
+import * as Rx from 'rxjs/Rx';
 
 import * as dotenv from 'dotenv';
 import * as vorpal from 'vorpal';
 
 dotenv.config();
 
-
-export const apiPath = process.env.API_URL;
-export const username = process.env.JUDGE_USERNAME;
-export const password = process.env.JUDGE_PASSWORD;
-export const fetchInterval = +process.env.FETCH_INTERVAL || 1000;
-
 export const vorpalBuilder = vorpal();
-
 
 vorpalBuilder
   .command('pull')
+  .option('-i, --interactive', 'Program acts interactively and if cannot decide about something, asks the user.')
   .description('Reads the database structure and creates entities on the selected output directory.')
   .action(function (args, callback) {
     this.log('Started judgings');
@@ -35,4 +29,5 @@ vorpalBuilder
 
 export const vorpalInstance = vorpalBuilder
   .delimiter('Entitype$')
-  .show();
+  .show()
+  .parse(process.argv);
